requirements.txt
python-telegram-bot==20.3
deep-translator==1.11.1
